Add files via upload
In this project, I applied supervised learning on hotel booking dataset that  will find a way to predict the chance of cancellation for a
particular hotel booking record. This was an originaly a Kaggle competiotion, can be found here, and was done as my Final-Project for the Data Mining cours under supervision of Dr. Ehsan Nazerfard.

As your first step, I
explored into the dataset which is formally called EDA. I determined
which features of the dataset are more informative. Then I took three methods (Decision Tree, Naïve Gaussian, and MLP) and
compared the results of them. I used accuracy as the metric and drew the relative confusion matrixs.
 This code was implemented by python
programming language and its related libraries like Scikit-learn, Tensorflow, Keras,
etc.